+++ packages/builders/radix-icons-builder/src/routes/+page.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import { Icon } from '@steeze-ui/svelte-icon'
-	import { Annotation } from '$lib'
+	import { ActivityLog } from '$lib/index.js'
 </script>
 
 <h1>Welcome to SvelteKit</h1>
 <p>Visit <a href="https://kit.svelte.dev">kit.svelte.dev</a> to read the documentation</p>
 
-<Icon src={Annotation} size="100" />
+<Icon src={ActivityLog} size="100" />