+++ packages/components/svelte-icon/src/lib/index.ts
@@ -1,1 +1,3 @@
+export type { IconSource } from './types.js'
+
 export { default as Icon } from './Icon.svelte'