+++ packages/components/svelte-icon/src/routes/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import Icon from '$lib/Icon.svelte'
-	import * as HeroIcons from '@steeze-ui/heroicons'
+	// import * as HeroIcons from '@steeze-ui/heroicons'
 	// import * as RadixIcons from '@steeze-ui/radix-icons'
 	// import * as IconicFreeIcons from '@steeze-ui/iconic-free'
 	// import * as Octicons from '@steeze-ui/octicons'
@@ -12,11 +12,11 @@
 	let octiconsTheme = 'default'
 
 	const iconpacks: any = {
-		Heroicons: {
-			icons: HeroIcons,
-			themes: ['default', 'solid'],
-			activeTheme: 'default'
-		}
+		// Heroicons: {
+		// 	icons: HeroIcons,
+		// 	themes: ['default', 'solid'],
+		// 	activeTheme: 'default'
+		// }
 		// RadixIcons: {
 		// 	icons: RadixIcons,
 		// 	themes: ['default'],