+++ packages/builders/remix-icons-builder/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @steeze-ui/radix-icons
 
+## 1.1.0
+
+### Features
+
+- update icon sources to version 3.3.0
+
 ## 1.0.0
 
 ### Major Changes